requirements.txt
autopep8==1.6.0
filemime==0.0.3
pycodestyle==2.9.1
python-magic==0.4.27
python-magic-bin==0.4.14
toml==0.10.2